fix: vuetify nuxt plugin template
